% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20041.R
\name{bubble_sort}
\alias{bubble_sort}
\title{Bubble sort method}
\usage{
bubble_sort(x)
}
\description{
Use bubble sort method to rank a vector
}
\examples{
\dontrun{
x=c(rep(1,100),rep(2,100),rep(0,100))
bubble_sort(x)
}
}
