% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20041.R
\name{choose_sort}
\alias{choose_sort}
\title{Choose sort method}
\usage{
choose_sort(x)
}
\description{
Use choose sort method to rank a vector
}
\examples{
\dontrun{
x=c(rep(1,100),rep(2,100),rep(0,100))
choose_sort(x)
}
}
