% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20041.R
\name{insert_sort}
\alias{insert_sort}
\title{Insert sort method}
\usage{
insert_sort(x)
}
\description{
Use insert sort method to rank a vector
}
\examples{
\dontrun{
x=c(rep(1,100),rep(2,100),rep(0,100))
insert_sort method(x)
}
}
