% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp20041.R
\name{quick_sort}
\alias{quick_sort}
\title{Quick sort method}
\usage{
quick_sort(x)
}
\description{
Use quick sort method to rank a vector
}
\examples{
\dontrun{
x=c(rep(1,100),rep(2,100),rep(0,100))
quick_sort(x)
}
}
