\name{StatComp20041-package}
\alias{StatComp20041-package}
\alias{StatComp20041}
\docType{package}
\title{
  Functions for ranking numbers
}
\description{
  This package provides several functions for ranking numbers from small to large
}

\author{
Ruxin Shi

Maintainer: Ruxin Shi <shiruxin@msu.edu>
}

\keyword{ package }


